config: add dtb-manifest for combined DTB image generation
Add a DTB manifest file defining the ordered set of device-tree blobs to be included in the combined DTB image for Qualcomm platforms.

The manifest currently covers:
- QCM6490 IDP
- QCS6490 RB3 Gen2
- Lemans EVK
- QCS9100 Ride (R0/R3)
- QCS8300 Ride
- Monaco EVK
- QCS615 Ride
- SA8775P Ride (R0/R3)
- Hamoa IoT EVK

This file is consumed as a configuration parameter by the combined DTB image tooling (e.g. build-dtb-image.sh in the image build pipeline) and provides a single source of truth for the DTB set and ordering used when generating dtb.bin for UEFI-based boot flows and CI builds.

Signed-off-by: Bjordis Collaku <bcollaku@qti.qualcomm.com>